rename ui files
